SONARJAVA-3849: Move S4838 test file to java-checks-test-sources
